switch from using 'group' attribute to using multiple webvtt files
This is probably a more flexible approach than over-defining a cue payload data structure. Spec and 'Raven' demo updated.
Browsers did not support multiple text tracks of kind=metadata until rather recently, and the group attribute had been a workaround for that, so now that it's not needed, it simplifies things a bit.